+++ dags/utils/shared_constants.py
